pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
Pillow-PIL = "^0.1dev"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.0"}
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"